pyproject.toml
[project]
name = "nerc_rates"
authors = [
  {name="MOC Alliance"},
]
version = "0.1"
readme = "README.md"
classifiers = [
  "Programming Language :: Python :: 3"
]
requires-python = ">=3.11"
dependencies = [
  "pydantic",
  "pyyaml",
  "requests",
]

[project.entry-points."console_scripts"]
validate-file = "nerc_rates.cmd.validate_file:main"


[build-system]
requires = [
  "setuptools>=42",
  "wheel"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"
